+++ libs/utils/src/services/blog/index.ts
@@ -3,3 +3,4 @@ export * from './list'
 export * from './like'
 export * from './paths'
 export * from './view'
+export * from './getBlogStaticProps'
